requirements.txt
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
asgiref==3.7.2
audioread==3.0.0
bcrypt==3.2.0
beautifulsoup4==4.6.3
black==23.3.0
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.0.7
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.2
duplicity==0.8.21
evdev==1.4.0
exceptiongroup==1.1.1
fasteners==0.14.1
ffprobe==0.5
fonttools==4.39.4
future==0.18.2
h11==0.14.0
html5lib==1.1
httpcore==0.17.2
httplib2==0.20.2
httpx==0.24.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
joblib==1.2.0
keyring==23.5.0
kiwisolver==1.4.4
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.2
librosa==0.10.0.post2
llvmlite==0.40.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.7.1
moddb==0.8.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numba==0.57.0
numpy==1.24.3
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
pathspec==0.11.1
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.6.0
praat-parselmouth==0.4.3
protobuf==3.12.4
psola==0.0.1
ptyprocess==0.7.0
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydub==0.25.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pypar==0.0.4
pyparsing==2.4.7
pyrate-limiter==2.8.1
pyRFC3339==1.1
pyTelegramBotAPI==4.11.0
pytest==7.3.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-decouple==3.8
python-magic==0.4.24
python-telegram-bot==20.3
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scikit-learn==1.2.2
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
sqlparse==0.4.4
systemd-python==234
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.11.2
tqdm==4.65.0
typing_extensions==4.6.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
zipp==1.0.0
